Conflicts:
	Server/pom.xml
	Server/src/main/java/com/inmobi/adserve/channels/server/ChannelServer.java
	Server/src/main/java/com/inmobi/adserve/channels/server/ChannelServerPipelineFactory.java
	Server/src/main/java/com/inmobi/adserve/channels/server/ChannelStatServerPipelineFactory.java
	Server/src/main/java/com/inmobi/adserve/channels/server/HttpRequestHandler.java
	Server/src/main/java/com/inmobi/adserve/channels/server/RequestIdHandler.java
	Server/src/main/java/com/inmobi/adserve/channels/server/ServletHandler.java
	Server/src/main/java/com/inmobi/adserve/channels/server/module/ServerModule.java
	Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/AsyncRequestMaker.java
	Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/Logging.java
	Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/MatchSegments.java
	Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/ResponseSender.java
	Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/filters/adgroup/impl/AdGroupDstFilter.java
	Server/src/main/java/com/inmobi/adserve/channels/server/servlet/ServletBackFill.java
	Server/src/main/java/com/inmobi/adserve/channels/server/servlet/ServletLogParser.java
	Server/src/test/java/com/inmobi/adserve/channels/server/ServerTest.java
	Server/src/test/java/com/inmobi/adserve/channels/server/requesthandler/filters/ChannelSegmentFilterApplierTest.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/adelphic/DCPAdelphicAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/ajillion/DCPAjillionAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/amobee/DCPAmobeeAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/appier/DCPAppierAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/appnexus/DCPAppNexusAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/drawbridge/DrawBridgeAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/generic/GenericAdapter.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/httpool/DCPHttPoolAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/huntmads/DCPHuntmadsAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/ifc/IFCAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/ifd/IFDAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/logan/DCPLoganAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/lomark/DCPLomarkAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/mable/DCPMableAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/madnet/DCPMadNetAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/mobfox/DCPMobFoxAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/mobilecommerce/MobileCommerceAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/mopub/DCPMoPubAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/mullahmedia/BaseMoolahMediaNetworkImpl.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/nexage/DCPNexageAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/openx/OpenxAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/paypal/DCPPayPalAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/placeiq/DCPPlaceIQAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/pubmatic/DCPPubmaticAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/siquis/DCPSiquisAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/smaato/DCPSmaatoAdnetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/tapit/DCPTapitAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/verve/DCPVerveAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/wapstart/DCPWapStartAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/webmoblink/WebmobLinkAdNetwork.java
	integration-adnetworks/src/main/java/com/inmobi/adserve/channels/adnetworks/xad/DCPxAdAdNetwork.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/AdNetworksTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPAdelphicAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPAjillionAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPAmobeeAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPAppNexusAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPHttpoolAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPHuntmadsAdNetworksTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPLoganAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPLomarkAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPMableAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPMadNetAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPMoPubAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPMobfoxAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPNexageAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPPlaceIQAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPPubmaticAdapterTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPSiquisTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPSmaatoAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPVerveAdapterTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPWapStartAdnetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPWiderPlanetAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPxAdAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/GenericAdapterTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/IFCAdNetworkTest.java
	integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/MullahMediaAdNetworkTest.java

+++ integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPAmobeeAdnetworkTest.java
@@ -74,7 +74,7 @@ public class DCPAmobeeAdnetworkTest extends TestCase {
         SASRequestParameters sasParams = new SASRequestParameters();
         CasInternalRequestParameters casInternalRequestParameters = new CasInternalRequestParameters();
         sasParams.setRemoteHostIp("206.29.182.240");
-        sasParams.setSlot("11");
+        sasParams.setSlot(Short.valueOf("11"));
         sasParams
                 .setUserAgent("Mozilla%2F5.0+%28iPhone%3B+CPU+iPhone+OS+5_0+like+Mac+OS+X%29+AppleWebKit%2F534.46+%28KHTML%2C+like+Gecko%29+Mobile%2F9A334");
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
@@ -161,7 +161,7 @@ public class DCPAmobeeAdnetworkTest extends TestCase {
         category.add(3l);
         sasParams.setCategories(category);
         casInternalRequestParameters.uid = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         String externalKey = "0344343";
         SlotSizeMapping.init();
         ChannelSegmentEntity entity = new ChannelSegmentEntity(AdNetworksTest.getChannelSegmentEntityBuilder(
@@ -188,7 +188,7 @@ public class DCPAmobeeAdnetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "38.5,-122.1514";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         casInternalRequestParameters.uid = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         List<Long> category = new ArrayList<Long>();
         category.add(3l);
         category.add(2l);
@@ -220,7 +220,7 @@ public class DCPAmobeeAdnetworkTest extends TestCase {
         casInternalRequestParameters.blockedCategories = new ArrayList<Long>(Arrays.asList(new Long[] { 50l, 51l }));
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         String externalKey = "19100";
         String beaconUrl = "http://c2.w.inmobi.com/c"
                 + ".asm/4/b/bx5/yaz/2/b/a5/m/0/0/0/202cb962ac59075b964b07152d234b70/4f8d98e2-4bbd"
@@ -247,7 +247,7 @@ public class DCPAmobeeAdnetworkTest extends TestCase {
         casInternalRequestParameters.blockedCategories = new ArrayList<Long>(Arrays.asList(new Long[] { 50l, 51l }));
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         sasParams.setSource("APP");
         String externalKey = "19100";
         String beaconUrl = "http://c2.w.inmobi.com/c"
