+++ integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/GenericAdapterTest.java
@@ -80,7 +80,7 @@ public class GenericAdapterTest extends TestCase {
         sasParams.setUserAgent("Mozilla");
         casInternalRequestParameters.uid = "1234";
         String externalKey = "118398";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         String beaconUrl = "http://c2.w.inmobi.com/c.asm/4/b/bx5/yaz/2/b/a5/m/0/0/0/202cb962ac59075b964b07152d234b70/4f8d98e2-4bbd-40bc-87e5-22da170600f9/-1/1/9cddca11?ds=1";
         ChannelSegmentEntity entity = new ChannelSegmentEntity(AdNetworksTest.getChannelSegmentEntityBuilder(
                 httpoolAdvertiserId, null, null, null, 0, null, null, true, true, externalKey, null, null, null, 0,
@@ -100,7 +100,7 @@ public class GenericAdapterTest extends TestCase {
         casInternalRequestParameters.uid = "1234";
         sasParams.setAdIncId(32);
         sasParams.setSiteIncId(18);
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         ChannelSegmentEntity entity = new ChannelSegmentEntity(AdNetworksTest.getChannelSegmentEntityBuilder(
                 httpoolAdvertiserId, null, null, null, 0, null, null, true, true, null, null, null, null, 0, true,
                 null, null, 0, null, false, false, false, false, false, false, false, false, false, false, null,
@@ -117,7 +117,7 @@ public class GenericAdapterTest extends TestCase {
         sasParams.setUserAgent("Mozilla");
         casInternalRequestParameters.uid = "1234";
         String externalKey = "118398";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         String beaconUrl = "http://c2.w.inmobi.com/c.asm/4/b/bx5/yaz/2/b/a5/m/0/0/0/202cb962ac59075b964b07152d234b70/4f8d98e2-4bbd-40bc-87e5-22da170600f9/-1/1/9cddca11?ds=1";
         ChannelSegmentEntity entity = new ChannelSegmentEntity(AdNetworksTest.getChannelSegmentEntityBuilder(
                 httpoolAdvertiserId, null, null, null, 0, null, null, true, true, externalKey, null, null, null, 0,
